% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{theme_tier4}
\alias{theme_tier4}
\title{Tier 4 default ggplot theme}
\usage{
theme_tier4(base_size = 12)
}
\value{
A ggplot2 theme object.
}
\description{
A SAFE-friendly theme with minimal gridlines and consistent
strips/legend.
}
